SteamReview/requirements.txt
flask == 1.1.2
flask-sqlalchemy
flask-script
flask_debugtoolbar
mysqlclient